Split things up so it's not quite so crazy
