+++ via/services/rewriter/__init__.py
@@ -1,0 +1,2 @@
+from via.services.rewriter.css import CSSRewriter
+from via.services.rewriter.html_lxml import LXMLRewriter
